requirements.txt
Django==6.0.2
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.9.0
gunicorn==23.0.0
Pillow==11.1.0
google-api-python-client==2.111.0
google-auth==2.27.0
